Moves physics components into component package
