+++ src/pyrite/_services/physics_service/__init__.py
@@ -12,8 +12,8 @@ from pyrite._types.service import ServiceProvider
 
 if TYPE_CHECKING:
     from collections.abc import Iterator, Sequence
-    from pyrite._physics.collider_component import ColliderComponent
-    from pyrite._physics.rigidbody_component import RigidbodyComponent
+    from pyrite._component.collider_component import ColliderComponent
+    from pyrite._component.rigidbody_component import RigidbodyComponent
     from pyrite._transform.transform import Transform
     from pyrite._types.constraint import Constraint
     from pyrite._types.shape import Shape
